# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Proker API",
        "description": "Dokumentasi API untuk Sistem Manajemen Program Kerja. Mencakup Authentication, Menu Management, Role & Group Management, dan Permission Management.",
        "contact": {
            "email": "admin@manajemenproker.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sys-menus": {
            "get": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Daftar Menu (Paginated)",
                "description": "Mengambil daftar seluruh menu sistem secara paginasi. Mendukung pencarian dan sorting.",
                "operationId": "eee2f25a23e1fc8ff38f833f5862c5f4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman aktif",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Urutan data (kolom asc/desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "page_size": {
                                            "type": "integer"
                                        },
                                        "total_page": {
                                            "type": "integer"
                                        },
                                        "total_records": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Tambah Menu Baru",
                "description": "Membuat menu baru dan otomatis menambahkan permission default (index, create, update, delete).",
                "operationId": "64f89b8ae903774db0f4ddcfcb72a15d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laporan"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/laporan"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Menu Laporan"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "📊"
                                    },
                                    "menu_order": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "visible": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "id_menu_parent": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sys-menus/tree": {
            "get": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Daftar Menu (Tree)",
                "description": "Mengambil daftar seluruh menu sistem dalam format tree (parent-children).",
                "operationId": "47bb5914a64f63a1f114a37466bff944",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_sys_menu": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "id_menu_parent": {
                                                        "type": "integer",
                                                        "nullable": true
                                                    },
                                                    "children": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sys-menus/{id}": {
            "get": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Detail Menu",
                "description": "Menampilkan detail satu menu berdasarkan ID.",
                "operationId": "89088e134027872c16041e2dd8a5bef6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Update Menu",
                "description": "Memperbarui data menu yang sudah ada.",
                "operationId": "4594aca0fc239d6819555c75137f624f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laporan Baru"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/laporan-baru"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "📈"
                                    },
                                    "menu_order": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "visible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sys Menu"
                ],
                "summary": "Hapus Menu",
                "description": "Menghapus menu beserta mapping permission yang terkait.",
                "operationId": "986da26190b7379b02b4b1d6dbecec3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {access_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}